Take a different approach to saving drafts.
Instead of trying to parse all the HTML of the draft
response, lets just assume the information on the
client is close enough and approximate the date. Then
we only need to parse out the messageId to be able
to update the draft later which should be less error
prone.

This should hopefully fix issue #24.