Cache the inbox list in localStorage.
This makes the inbox list not take forever to load in
the common case of going back, but it does mean you
need to click 'Refresh' often when the page first
loads. It is smart enough to not ask you to refresh
if the list is visually identical and the only difference
if the dates.

I fear this is going to be super annoying to click
'Refresh' so much, but lets try it out.

Fixes #8.