Remove click to refresh for the issue inbox.
The proved to be too annoying and weird. Most of the
time the issues that changed aren't even the ones at
the top you're about to click on. Rietveld is also so
slow that it usually takes 4 seconds for the issues
to change out from under you.

Lets just always refresh the issues once they load.

Fixes #37.