Allow pressing the 'm' key to open the publish dialog.
This patch fixes keyboard handling for letter keys in <cr-keyboard>
and adds support for global key bindings that are only active
when no editable element is focused and no dialogs are open.

It then uses this new support to add a handler for the 'm' key that
opens the publish dialog. I also removed the top: 100px from
common.css so that the dialog is vertically positioned at your scroll
offset so pressing 'm' when scrolled way down a patch brings up the
dialog in the center of your screen. Note that the polyfill dialog
does not support vertical centering so it'll jump to the top.

This is the first part of issue #3.