Use a map of file extensions instead of an if tree.
The if tree and regex matching is very expensive on large
patchsets. Using a map reduces the time spent creating the
PatchFile objects by 50% in large rebaseline patches where
there's a ton of files that don't map to any syntax and
would hit every case in the old if tree.

ex. https://codereview.chromium.org/328053002 is reduced by
100ms when loading the really big patchset.